# perf(ext/urlpattern): optimize URLPattern.exec