@bartlomieju 2023-08-15
Asking @littledivy for review who has a WIP patch that optimizes `URLPattern` to use JS implementation.